Introduce endpoint:: Michel Namespace (#48)
Move all Michels into namespace endpoint::, in advance of upcoming trackless::.

Introduce dummy trackless:: namespace with trackless::MichelEvent and trackless::GetQualityMichels.

Move various endpoint michel functions (CompareMichels and GetQualityMichels) into Michel.h.

Completely reproduces previous result aka backwards compatible and nothing actually changed yet.
